django/requirements.txt
Django==4.0.2
psycopg2-binary==2.9.1
django-environ
pytz==2022.2.1
djangorestframework
djangorestframework-simplejwt
